+++ src/main/java/com/example/OrderCoffeeBE/Service/OrderService.java
@@ -9,8 +9,8 @@ import java.util.List;
 public interface OrderService {
     List<Order> findAll();
     Order createOrder(PostOrderDTO orderDTO);
-    Order updateOrder(int id, OrderDTO orderDTO);
+    Order updateOrder(Long id, OrderDTO orderDTO);
     void sortDeleteOrder(int id);
-    Order findById(int id);
+    Order findById(long id);
 
 }
