+++ src/main/java/com/example/OrderCoffeeBE/repository/ProductRepository.java
@@ -3,6 +3,6 @@ package com.example.OrderCoffeeBE.repository;
 import com.example.OrderCoffeeBE.Model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> { ;
+public interface ProductRepository extends JpaRepository<Product, Long> { ;
     Product findByName(String name);
 }
